+++ src/core_modules/capture-core/components/WidgetEnrollmentEventNew/WidgetEnrollmentEventNew.types.js
@@ -1,5 +1,5 @@
 // @flow
-import type { ExternalSaveHandler } from './common.types';
+import type { ExternalSaveHandler, RulesExecutionDependencies } from './common.types';
 
 export type WidgetProps = {|
     programId: string,
@@ -7,6 +7,7 @@ export type WidgetProps = {|
     teiId: string,
     enrollmentId: string,
     orgUnitId: string,
+    rulesExecutionDependencies: RulesExecutionDependencies,
     onSave?: ExternalSaveHandler,
     onSaveSuccessActionType?: string,
     onSaveErrorActionType?: string,